+++ sync-for-payables/src/codatsyncpayables/models/operations/delete_bill_attachment.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import errormessage as shared_errormessage
-from ..shared import pushoperation as shared_pushoperation
+from ...models.shared import pushoperation as shared_pushoperation
 from typing import Optional
 
 
@@ -26,13 +25,11 @@ class DeleteBillAttachmentRequest:
 class DeleteBillAttachmentResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
     push_operation: Optional[shared_pushoperation.PushOperation] = dataclasses.field(default=None)
     r"""OK"""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""