+++ sync-for-payables/src/codatsyncpayables/models/operations/get_accounting_profile.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import companyinfo as shared_companyinfo
-from ..shared import errormessage as shared_errormessage
+from ...models.shared import companyinfo as shared_companyinfo
 from typing import Optional
 
 
@@ -20,13 +19,11 @@ class GetAccountingProfileRequest:
 class GetAccountingProfileResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
     company_info: Optional[shared_companyinfo.CompanyInfo] = dataclasses.field(default=None)
     r"""Success"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""