+++ sync-for-payables/src/codatsyncpayables/models/operations/get_tracking_category.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import errormessage as shared_errormessage
-from ..shared import trackingcategorytree as shared_trackingcategorytree
+from ...models.shared import trackingcategorytree as shared_trackingcategorytree
 from typing import Optional
 
 
@@ -22,12 +21,10 @@ class GetTrackingCategoryRequest:
 class GetTrackingCategoryResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""
     tracking_category_tree: Optional[shared_trackingcategorytree.TrackingCategoryTree] = dataclasses.field(default=None)
     r"""Success"""